fix: dashboard disconnect on refresh tab
