+++ hooks/use-autoclaim.ts
@@ -2,11 +2,9 @@ import { useCallback, useEffect, useState } from "react";
 import { useAccount, useWriteContract } from "wagmi";
 import CONTRACTS from "@/utils/contracts";
 import claimRegistryABI from "@/utils/abis/claimRegistry";
-import { Address } from "viem";
-import { GetPublicClientReturnType, getPublicClient } from "wagmi/actions";
+import { getPublicClient } from "wagmi/actions";
 import { config } from "@/wagmi";
-
-const BLOCK_RANGE_SIZE = 50000;
+import { fetchRegister, fetchUnregister } from "@/utils/fetchEvents";
 
 function useAutoclaim() {
   const account = useAccount();
@@ -67,65 +65,6 @@ function useAutoclaim() {
   }, [account]);
 
   return { register, updateConfig, unregister, isRegister };
-}
-
-async function fetchRegister(claimRegistryAddress: Address, userAddress: Address, fromBlock: bigint, client: GetPublicClientReturnType) {
-  if (!client) return [];
-  const toBlock = await client.getBlockNumber();
-
-  let allEvents = [];
-
-  while (fromBlock <= toBlock) {
-    const nextBlock = fromBlock + BigInt(BLOCK_RANGE_SIZE) > toBlock ? toBlock : fromBlock + BigInt(BLOCK_RANGE_SIZE);
-
-    console.log(`Fetching from block ${fromBlock} to ${nextBlock}`);
-
-    const events = await client.getContractEvents({
-      abi: claimRegistryABI,
-      address: claimRegistryAddress,
-      eventName: "Register",
-      args: {
-        user: userAddress,
-      },
-      fromBlock: fromBlock,
-      toBlock: nextBlock,
-    });
-
-    allEvents.push(...events);
-    fromBlock = nextBlock + BigInt(1);
-  }
-
-  return allEvents;
-}
-
-async function fetchUnregister(claimRegistryAddress: Address, userAddress: Address, fromBlock: bigint, client: GetPublicClientReturnType) {
-  if (!client) return [];
-  const toBlock = await client.getBlockNumber();
-
-  let currentBlock = BigInt(fromBlock);
-  let allEvents = [];
-
-  while (currentBlock <= toBlock) {
-    const nextBlock = currentBlock + BigInt(BLOCK_RANGE_SIZE) > toBlock ? toBlock : currentBlock + BigInt(BLOCK_RANGE_SIZE);
-
-    console.log(`Fetching from block ${currentBlock} to ${nextBlock}`);
-
-    const events = await client.getContractEvents({
-      abi: claimRegistryABI,
-      address: claimRegistryAddress,
-      eventName: "Unregister",
-      args: {
-        user: userAddress,
-      },
-      fromBlock: currentBlock,
-      toBlock: nextBlock,
-    });
-
-    allEvents.push(...events);
-    currentBlock = nextBlock + BigInt(1);
-  }
-
-  return allEvents;
 }
 
 export default useAutoclaim;
